// tests/regression/test/issue-2411/schema.ts
//////////////////////////////////////////////////////////////////////////////////////////////
// DO NOT MODIFY THIS FILE                                                                  //
// This file is automatically generated by ZenStack CLI and should not be manually updated. //
//////////////////////////////////////////////////////////////////////////////////////////////

/* eslint-disable */

import { type SchemaDef, type AttributeApplication, type FieldDefault, ExpressionUtils } from "@zenstackhq/schema";
export class SchemaType implements SchemaDef {
    provider = {
        type: "sqlite"
    } as const;
    models = {
        Foo: {
            name: "Foo",
            fields: {
                id: {
                    name: "id",
                    type: "String",
                    id: true,
                    attributes: [{ name: "@id" }, { name: "@default", args: [{ name: "value", value: ExpressionUtils.call("cuid") }] }] as readonly AttributeApplication[],
                    default: ExpressionUtils.call("cuid") as FieldDefault
                },
                createdAt: {
                    name: "createdAt",
                    type: "DateTime",
                    attributes: [{ name: "@default", args: [{ name: "value", value: ExpressionUtils.call("now") }] }] as readonly AttributeApplication[],
                    default: ExpressionUtils.call("now") as FieldDefault
                },
                updatedAt: {
                    name: "updatedAt",
                    type: "DateTime",
                    updatedAt: true,
                    attributes: [{ name: "@updatedAt" }] as readonly AttributeApplication[]
                },
                metadata: {
                    name: "metadata",
                    type: "Metadata",
                    attributes: [{ name: "@json" }] as readonly AttributeApplication[]
                },
                optionalMetadata: {
                    name: "optionalMetadata",
                    type: "Metadata",
                    optional: true,
                    attributes: [{ name: "@json" }] as readonly AttributeApplication[]
                }
            },
            idFields: ["id"],
            uniqueFields: {
                id: { type: "String" }
            }
        }
    } as const;
    typeDefs = {
        Metadata: {
            name: "Metadata",
            fields: {
                someString: {
                    name: "someString",
                    type: "String"
                },
                someInt: {
                    name: "someInt",
                    type: "Int"
                }
            }
        }
    } as const;
    plugins = {};
}
export const schema = new SchemaType();